docker/consumer-stream/requirements.txt
pandas
psycopg2-binary
kafka-python
dotenv
